refactor: 도메인 별로 타입 정의 분리
