+++ src/hooks/queries.ts
@@ -4,32 +4,33 @@ import {
   useQueryClient,
   useInfiniteQuery,
 } from '@tanstack/react-query';
-import { apiService } from '@/lib/api';
+import { themesApi, productsApi, authApi, ordersApi } from '@/api';
 import type { GiftOrderForm } from '@/types';
 
 export const queries = {
   themes: {
     key: ['themes'] as const,
-    fn: () => apiService.getThemes(),
+    fn: () => themesApi.getThemes(),
   },
   themeInfo: {
     key: (themeId: string | number) => ['themes', themeId, 'info'] as const,
-    fn: (themeId: string | number) => apiService.getThemeInfo(themeId),
+    fn: (themeId: string | number) => themesApi.getThemeInfo(themeId),
   },
   themeProducts: {
     key: (themeId: string | number) => ['themes', themeId, 'products'] as const,
-    fn: (themeId: string | number) => apiService.getThemeProducts(themeId),
+    fn: (themeId: string | number) => themesApi.getThemeProducts(themeId),
   },
   rankingProducts: {
     key: (targetType: string, rankType: string) =>
       ['products', 'ranking', targetType, rankType] as const,
     fn: (targetType: string, rankType: string) =>
-      apiService.getRankingProducts(targetType, rankType),
+      productsApi.getRankingProducts(targetType, rankType),
   },
   productSummary: {
     key: (productId: string | number) =>
       ['products', productId, 'summary'] as const,
-    fn: (productId: string | number) => apiService.getProductSummary(productId),
+    fn: (productId: string | number) =>
+      productsApi.getProductSummary(productId),
   },
   products: {
     key: ['products'] as const,
@@ -82,11 +83,11 @@ export function useThemeProductsInfiniteQuery(
   return useInfiniteQuery({
     queryKey: [...queries.themeProducts.key(themeId), 'infinite'],
     queryFn: ({ pageParam = 0 }) =>
-      apiService.getThemeProductsWithPagination(themeId, {
+      themesApi.getThemeProductsWithPagination(themeId, {
         cursor: pageParam,
         limit,
       }),
-    getNextPageParam: (lastPage, allPages) => {
+    getNextPageParam: (lastPage: any, allPages) => {
       const currentCursor = allPages.length * limit;
       return lastPage.hasMoreList ? currentCursor : undefined;
     },
@@ -99,7 +100,7 @@ export function useLoginMutation() {
 
   return useMutation({
     mutationFn: (credentials: { email: string; password: string }) =>
-      apiService.login(credentials),
+      authApi.login(credentials),
     onSuccess: data => {
       sessionStorage.setItem('userInfo', JSON.stringify(data));
 
@@ -112,7 +113,7 @@ export function useCreateOrderMutation() {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (orderData: GiftOrderForm) => apiService.createOrder(orderData),
+    mutationFn: (orderData: GiftOrderForm) => ordersApi.createOrder(orderData),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: queries.products.key });
     },
