+++ src/types/api.ts
@@ -1,0 +1,9 @@
+import type { Product } from './product';
+
+export interface PaginatedResponse<T> {
+  data: T[];
+  hasMoreList: boolean;
+  nextCursor?: number;
+}
+
+export type ThemeProductsResponse = PaginatedResponse<Product>;
