+++ src/types/theme.ts
@@ -1,0 +1,20 @@
+export interface GiftTheme {
+  themeId: number;
+  name: string;
+  image: string;
+}
+
+export interface ThemeInfo {
+  themeId: number;
+  name: string;
+  title: string;
+  description: string;
+  backgroundColor: string;
+}
+
+export interface MessageCardTemplate {
+  id: number;
+  thumbUrl: string;
+  imageUrl: string;
+  defaultTextMessage: string;
+}
